backend/requirements.txt
fastapi
sqlalchemy
asyncpg
sqlalchemy[asyncio]
python-dotenv
pydantic-settings
alembic

bcrypt<4.0.0
passlib[bcrypt]
python-jose[cryptography]
uvicorn